refactor: rename to `prevent-push-to-default-branch`
